+++ demos/browser/app/meetingV2/meetingV2.ts
@@ -148,7 +148,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
 
   // feature flags
   enableWebAudio = false;
-  enableUnifiedPlanForChromiumBasedBrowsers = false;
+  enableUnifiedPlanForChromiumBasedBrowsers = true;
   enableSimulcast = false;
 
   markdown = require('markdown-it')({linkify: true});
@@ -317,22 +317,17 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
       const collections = optionalFeatures.selectedOptions;
       this.enableSimulcast = false;
       this.enableWebAudio = false;
-      this.enableUnifiedPlanForChromiumBasedBrowsers = false;
+      this.enableUnifiedPlanForChromiumBasedBrowsers = true;
       this.log("Feature lists:");
       for (let i = 0; i < collections.length; i++) {
         // hard code magic
         if (collections[i].value === 'simulcast') {
           this.enableSimulcast = true;
-          this.enableUnifiedPlanForChromiumBasedBrowsers = true;
           this.log('attempt to enable simulcast');
         }
         if (collections[i].value === 'webaudio') {
           this.enableWebAudio = true;
           this.log('attempt to enable webaudio');
-        }
-        if (collections[i].value === 'unifiedplan') {
-          this.enableUnifiedPlanForChromiumBasedBrowsers = true;
-          this.log('attempt to enable unified plan');
         }
       }
     });